+++ myWidget/Views/WidgetMediumCardView.swift
@@ -1,5 +1,5 @@
 //
-//  widgetLargeCardView.swift
+//  SwiftUIView.swift
 //  포스밀
 //
 //  Created by Kim Insub on 2022/07/23.
@@ -8,30 +8,34 @@
 import SwiftUI
 import WidgetKit
 
-struct widgetLargeCardView: View{
-    @Environment(\.colorScheme) var colorScheme
+struct WidgetMediumCardView: View {
+
     @State var LUNCH_FOODS: [String]? = []
     @State var DINNER_FOODS: [String]? = []
-    let CURRENTDATE: [String:String]?
-    let scheduleDict: [String:String]
-    let mealNameDict: [String:String]
 
-    var body: some View{
+//    let scheduleDict: [String:String]
+//    let mealNameDict: [String:String]
 
-        VStack{
+    let scheduleDict = TimeManager().schedule
+    let mealNameDict = TimeManager().mealNameDict
+
+    var body: some View {
+
+        HStack{
             lunchMenuViewBuilder()
             Divider()
-                .padding(.horizontal)
+                .padding(.vertical)
             dinnerMenuViewBuilder()
         }
-        .background(Color.cardBackground)
         .onAppear{
             self.LUNCH_FOODS = []
-            self.LUNCH_FOODS = UserDefaults(suiteName: "group.com.kim.widgetProject")!.stringArray(forKey: "LUNCH") as [String]?
+            self.LUNCH_FOODS = UserDefaults(suiteName: "group.com.kim.widgetProject")?.stringArray(forKey: "LUNCH") as? [String]
 
             self.DINNER_FOODS = []
-            self.DINNER_FOODS = UserDefaults(suiteName: "group.com.kim.widgetProject")!.stringArray(forKey: "DINNER") as [String]?
+            self.DINNER_FOODS = UserDefaults(suiteName: "group.com.kim.widgetProject")?.stringArray(forKey: "DINNER") as? [String]
         }
+        .background(Color.cardBackground)
+
     }
 
     @ViewBuilder func lunchMenuViewBuilder() -> some View {
@@ -97,6 +101,7 @@ struct widgetLargeCardView: View{
                     .background(Color.timeBackgroundColor)
                     .cornerRadius(10)
             }
+
             //Contents
             VStack(alignment: .leading, spacing: 2){
                 if DINNER_FOODS != nil{
@@ -122,6 +127,5 @@ struct widgetLargeCardView: View{
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
         .padding()
-
     }
 }
